Fix bug involving order of loading params from URL
If the data file is loaded last, then the hit index will be set to zero since
there are no hits to display until the data file is loaded. So I rearranged the
order of the controllerMap entries so that the data file is first. In parseURL,
the keys are drawn from controllerMap instead of metadata to take advantage of
the new ordering, and because those are the only relevant keys anyways.